+++ src/loading-state/main.js
@@ -1,5 +1,5 @@
-import axios from 'axios';
 import { create as createUseAsync } from '@postinumero/use-async/loading-state';
+import axios from 'axios';
 
 export const create = (config) =>
   createUseAsync(typeof config === 'function' ? config : axios.create(config));